Refactor - Use instances, not classes
We were using classes to specify the strategy to be used for picking
the next router. As each strategy has both infrastructural and runtime
dependencies, like the least busy agent picker had a dependency on
neutron client, single agent picker had a dependency on a user
configuration value - the agent to be picked each time. We injected
neutron client at runtime, but the parameter at configuration
time. This looked wrong that some of the dependencies were class
attributes, others were instance attributes. This change moves all the
dependencies to instance attributes.