Refactor: extract SSH and remote command execution
Separate out the responsibilities for connecting to a remote host, and
create the Host abstraction that could run commands on a host. Also add
some test helpers to ease creating a mock ssh connection instance. With
this change, the responsibilites of RemoteNodeCleanup became smaller,
and ssh related code could be re-used later. Also the internal
ssh_client is hidden, so RemoteLbaasV2Cleanup does not need to know how
the connection happens